feat: adjusted database structure
